recipe 9-5 start

+++ chapter9/recipe9-5/start/football/src/main/java/com/packt/football/service/DynamicQueriesService.java
@@ -112,9 +112,9 @@ public class DynamicQueriesService {
                 .collect(Collectors.toList());
     }
 
-    public BigInteger countPlayers() {
+    public Long countPlayers() {
         Query query = em.createNativeQuery("SELECT COUNT(1) FROM players");
-        return (BigInteger) query.getSingleResult();
+        return (Long) query.getSingleResult();
     }
 
     public Player findPlayerById(Integer id) {
